requirements.txt
amqp==2.1.4
beautifulsoup4==4.7.1
billiard==3.5.0.2
boto==2.46.1
boto3==1.4.4
botocore==1.5.95
bs4==0.0.1
celery==4.1.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==1.11.4
django-allauth==0.32.0
django-braces==1.13.0
django-cacheback==1.3.2
django-cors-headers==1.1.0
django-crontab==0.7.1
django-debug-toolbar==1.5
django-extensions==1.7.4
django-filter==1.0.4
django-model-utils==2.6
django-mysql==1.1.0
django-oauth-toolkit==0.10.0
django-redis==4.7.0
django-rest-swagger==2.1.2
djangorestframework==3.6.4
djangorestframework-jwt==1.10.0
docutils==0.14
elasticsearch==2.3.0
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
kombu==4.0.2
Markdown==2.6.7
MarkupSafe==1.1.0
mysqlclient==1.3.13
oauthlib==1.0.3
olefile==0.46
openapi-codec==1.3.2
Pillow==4.1.1
psycopg2==2.7.1
pycurl==7.43.0.2
PyJWT==1.7.1
pymongo==3.7.2
python-dateutil==2.7.5
python3-openid==3.1.0
pytz==2017.2
redis==3.0.1
requests==2.13.0
requests-oauthlib==1.0.0
s3transfer==0.1.13
simplejson==3.16.0
six==1.12.0
soupsieve==1.7.3
SQLAlchemy==1.2.16
sqlparse==0.2.4
tornado==5.1.1
uritemplate==3.0.0
urllib3==1.24.1
uWSGI==2.0.17.1
vine==1.1.3
